autogenerate type map
