+++ src/snarky/gen/transaction-json.ts
@@ -10,11 +10,23 @@ import {
   AuthRequired,
   Sign,
   AuthorizationKind,
-  TypeMap,
 } from '../transaction-leaves-json.js';
 
 export { ZkappCommand, AccountUpdate };
 export * from '../transaction-leaves-json.js';
+export { TypeMap };
+
+type TypeMap = {
+  PublicKey: PublicKey;
+  UInt64: UInt64;
+  UInt32: UInt32;
+  TokenId: TokenId;
+  Field: Field;
+  Bool: Bool;
+  AuthRequired: AuthRequired;
+  Sign: Sign;
+  AuthorizationKind: AuthorizationKind;
+};
 
 type ZkappCommand = {
   feePayer: {
