+++ src/snarky/transaction-leaves-json.ts
@@ -10,8 +10,6 @@ export {
   TokenId,
 };
 
-export { TypeMap };
-
 type Field = string;
 type Bool = boolean;
 type UInt64 = string;
@@ -21,16 +19,3 @@ type Sign = 'Positive' | 'Negative';
 type AuthRequired = 'Signature' | 'Proof' | 'Either' | 'None' | 'Impossible';
 type AuthorizationKind = 'Signature' | 'Proof' | 'None_given';
 type TokenId = Field;
-
-// to what types in the js layout are mapped
-type TypeMap = {
-  PublicKey: PublicKey;
-  Field: Field;
-  Bool: Bool;
-  AuthRequired: AuthRequired;
-  AuthorizationKind: AuthorizationKind;
-  UInt32: UInt32;
-  UInt64: UInt64;
-  Sign: Sign;
-  TokenId: TokenId;
-};
